+++ tests/testthat/test-deprecated.R
@@ -4,7 +4,6 @@ test_that("default deprecations behave as expected", {
   on.exit(env_unbind(deprecation_env, "test"))
   local_options(lifecycle_verbosity = "default")
 
-  expect_condition(deprecate_soft("1.0.0", "foo()"), class = "lifecycle_soft_deprecated")
   expect_warning(deprecate_warn("1.0.0", "foo()", id = "test"), class = "lifecycle_warning_deprecated")
   expect_warning(deprecate_warn("1.0.0", "foo()", id = "test"), NA)
   expect_defunct(deprecate_stop("1.0.0", "foo()"))
@@ -40,7 +39,6 @@ test_that("quiet suppresses _soft and _warn", {
 test_that("warning always warns in _soft and _warn", {
   local_options(lifecycle_verbosity = "warning")
 
-  expect_deprecated(deprecate_soft("1.0.0", "foo()"))
   expect_deprecated(deprecate_warn("1.0.0", "foo()"))
   expect_defunct(deprecate_stop("1.0.0", "foo()"))
 })
@@ -54,7 +52,7 @@ test_that("error coverts _soft and _warn to errors", {
 })
 
 test_that("soft deprecation uses correct calling envs", {
-  local_options(lifecycle_verbosity = "default")
+  withr::local_envvar(TESTTHAT_PKG = "testpackage")
 
   # Simulate package functions available from global environment
   env <- new_environment(parent = ns_env("lifecycle"))
@@ -68,15 +66,14 @@ test_that("soft deprecation uses correct calling envs", {
   })
   local_bindings(!!!as.list(env), .env = global_env())
 
-  # Calling package function directly should warning
+  # Calling package function directly should warn
   cnd <- catch_cnd(evalq(softly(), global_env()), "warning")
   expect_s3_class(cnd, class = "lifecycle_warning_deprecated")
   expect_match(conditionMessage(cnd), "lifecycle")
 
   # Calling package function indirectly from global env shouldn't
-  cnd <- catch_cnd(evalq(softly_softly(), global_env()), "lifecycle_soft_deprecated")
-  expect_s3_class(cnd, class = "lifecycle_soft_deprecated")
-  expect_match(conditionMessage(cnd), "lifecycle")
+  cnd <- catch_cnd(evalq(softly_softly(), global_env()), "warning")
+  expect_equal(cnd, NULL)
 })
 
 test_that("warning conditions are signaled only once if warnings are suppressed", {